+++ src/routes/posts/+server.js
@@ -1,0 +1,7 @@
+import posts from '../../posts.json';
+
+export const GET = async () => {
+  return new Response(JSON.stringify(posts));
+};
+
+
